documents-to-corpus/requirements.txt
nltk==3.4.5
progressbar2==3.47.0